feat: show collection items in collection layout page
