+++ src/common/ArticlePageHeader.ts
@@ -1,10 +1,11 @@
 import type BreadcrumbProps from "./Breadcrumb"
+import type CollectionCardProps from "./CollectionCard"
 
 export interface ArticlePageHeaderProps {
   breadcrumb: BreadcrumbProps
-  category: string
   title: string
-  date: string
+  category: CollectionCardProps["category"]
+  date: CollectionCardProps["lastUpdated"]
   summary: string[]
   LinkComponent?: any
 }
