+++ src/engine/metadata.ts
@@ -1,4 +1,5 @@
-import type { IsomerPageSchema, IsomerSitemap } from "./types"
+import { getSitemapAsArray } from "~/utils"
+import type { IsomerPageSchema, IsomerSitemap } from "~/types"
 
 type SitemapXmlItem = {
   url: string
@@ -10,7 +11,7 @@ export const getMetadata = (props: IsomerPageSchema) => {
     metadataBase: props.site.url ? new URL(props.site.url) : undefined,
     description: props.page.description || undefined,
     robots: {
-      index: !props.page.noIndex,
+      index: !(props.page.type === "page" && props.page.noIndex),
     },
     icons: {
       shortcut:
@@ -22,7 +23,7 @@ export const getMetadata = (props: IsomerPageSchema) => {
     },
   }
 
-  if (props.page.permalink === "/") {
+  if (props.page.type === "page" && props.page.permalink === "/") {
     return metadata
   }
 
@@ -53,20 +54,12 @@ export const getRobotsTxt = (props: IsomerPageSchema) => {
 }
 
 export const getSitemapXml = (sitemap: IsomerSitemap) => {
-  let result: SitemapXmlItem[] = []
-
-  const traverse = (node: IsomerSitemap) => {
-    if (node.permalink) {
-      result.push({
-        url: node.permalink,
-        lastModified: node.lastModified,
-      })
-    }
-    if (node.children && node.children.length > 0) {
-      node.children.forEach((child) => traverse(child))
-    }
-  }
-
-  traverse(sitemap)
-  return result
+  return getSitemapAsArray(sitemap)
+    .filter((item) => item.type === "page")
+    .map(
+      ({ permalink, lastModified }): SitemapXmlItem => ({
+        url: permalink,
+        lastModified,
+      }),
+    )
 }
