+++ src/types/index.ts
@@ -1,0 +1,5 @@
+export type { IsomerComponent } from "./components"
+export type * from "./page"
+export type * from "./schema"
+export type { IsomerSiteProps } from "./site"
+export type { IsomerSitemap } from "./sitemap"
