+++ src/utils/getSitemapAsArray.ts
@@ -1,0 +1,18 @@
+import type { IsomerSitemap } from "~/engine"
+
+export const getSitemapAsArray = (sitemap: IsomerSitemap) => {
+  let result: IsomerSitemap[] = []
+
+  const traverse = (node: IsomerSitemap) => {
+    if (node.permalink) {
+      const { children, ...rest } = node
+      result.push(rest)
+    }
+    if (node.children && node.children.length > 0) {
+      node.children.forEach((child) => traverse(child))
+    }
+  }
+
+  traverse(sitemap)
+  return result
+}
