+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export { getBreadcrumbFromSiteMap } from "./getBreadcrumbFromSiteMap"
 export { getSanitizedInlineContent } from "./getSanitizedInlineContent"
+export { getSitemapAsArray } from "./getSitemapAsArray"
